+++ source/full/components/pages/DisabledDomainsPage.tsx
@@ -48,7 +48,7 @@ export function DisabledDomainsPage() {
             getToaster().show({
                 intent: Intent.DANGER,
                 message: t("error.generic", { message: localisedErrorMessage(err) }),
-                timeout: 10000
+                timeout: 5000
             });
         }
         setRemoveDomain(null);