+++ source/full/components/pages/NotificationsPage.tsx
@@ -29,7 +29,7 @@ export function NotificationsPage() {
             getToaster().show({
                 intent: Intent.DANGER,
                 message: t("error.generic", { message: localisedErrorMessage(err) }),
-                timeout: 10000
+                timeout: 5000
             });
         });
     }, []);