+++ source/popup/components/pages/SettingsPage.tsx
@@ -71,7 +71,7 @@ export function SettingsPage() {
             getToaster().show({
                 intent: Intent.DANGER,
                 message: t("error.generic", { message: localisedErrorMessage(err) }),
-                timeout: 10000
+                timeout: 5000
             });
         }
     }, []);
@@ -83,7 +83,7 @@ export function SettingsPage() {
             getToaster().show({
                 intent: Intent.DANGER,
                 message: t("error.generic", { message: localisedErrorMessage(err) }),
-                timeout: 10000
+                timeout: 5000
             });
         }
     }, []);
@@ -94,7 +94,7 @@ export function SettingsPage() {
             getToaster().show({
                 intent: Intent.DANGER,
                 message: t("error.reset", { message: localisedErrorMessage(err) }),
-                timeout: 10000
+                timeout: 5000
             });
         });
     }, []);